Union of the domain and web API projects
