--- TradingSystem/TradingSystem.WebApi/TradingSystem.WebApi.csproj
@@ -1,17 +1,0 @@
-<Project Sdk="Microsoft.NET.Sdk.Web">
-
-  <PropertyGroup>
-    <TargetFramework>netcoreapp3.1</TargetFramework>
-    <UserSecretsId>78e13616-b6c6-48e4-90c7-d84b378cab89</UserSecretsId>
-  </PropertyGroup>
-
-  <ItemGroup>
-    <PackageReference Include="Microsoft.AspNetCore.Cors" Version="2.2.0" />
-  </ItemGroup>
-
-  <ItemGroup>
-    <ProjectReference Include="..\TradingSystem\TradingSystem.csproj" />
-  </ItemGroup>
-
-
-</Project>
