+++ TradingSystem/TradingSystem.sln
@@ -3,8 +3,6 @@ Microsoft Visual Studio Solution File, Format Version 12.00
 # Visual Studio Version 16
 VisualStudioVersion = 16.0.31112.23
 MinimumVisualStudioVersion = 10.0.40219.1
-Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "TradingSystem", "TradingSystem\TradingSystem.csproj", "{F14D71E5-9DE5-48E9-B331-C156F5ACE0DD}"
-EndProject
 Project("{2150E333-8FDC-42A3-9474-1A3956D46DE8}") = "Solution Items", "Solution Items", "{2C47C7B7-5B09-46B9-8EDD-FBC6756B49CC}"
 	ProjectSection(SolutionItems) = preProject
 		.editorconfig = .editorconfig
@@ -14,7 +12,7 @@ Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "TradingSystemTests", "Tradi
 EndProject
 Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "AcceptanceTests", "AcceptanceTests\AcceptanceTests.csproj", "{3EFAE6AF-DB02-40FC-BD8E-B11305970B98}"
 EndProject
-Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "TradingSystem.WebApi", "TradingSystem.WebApi\TradingSystem.WebApi.csproj", "{C81E9F00-7899-4097-A7CE-043E818BE0E6}"
+Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "TradingSystem", "TradingSystem\TradingSystem.csproj", "{09079C11-2362-4FDE-ABD8-8309B75B55AE}"
 EndProject
 Global
 	GlobalSection(SolutionConfigurationPlatforms) = preSolution
@@ -22,10 +20,6 @@ Global
 		Release|Any CPU = Release|Any CPU
 	EndGlobalSection
 	GlobalSection(ProjectConfigurationPlatforms) = postSolution
-		{F14D71E5-9DE5-48E9-B331-C156F5ACE0DD}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
-		{F14D71E5-9DE5-48E9-B331-C156F5ACE0DD}.Debug|Any CPU.Build.0 = Debug|Any CPU
-		{F14D71E5-9DE5-48E9-B331-C156F5ACE0DD}.Release|Any CPU.ActiveCfg = Release|Any CPU
-		{F14D71E5-9DE5-48E9-B331-C156F5ACE0DD}.Release|Any CPU.Build.0 = Release|Any CPU
 		{828D39F7-185D-44DC-954E-1076CD81E23F}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
 		{828D39F7-185D-44DC-954E-1076CD81E23F}.Debug|Any CPU.Build.0 = Debug|Any CPU
 		{828D39F7-185D-44DC-954E-1076CD81E23F}.Release|Any CPU.ActiveCfg = Release|Any CPU
@@ -34,10 +28,10 @@ Global
 		{3EFAE6AF-DB02-40FC-BD8E-B11305970B98}.Debug|Any CPU.Build.0 = Debug|Any CPU
 		{3EFAE6AF-DB02-40FC-BD8E-B11305970B98}.Release|Any CPU.ActiveCfg = Release|Any CPU
 		{3EFAE6AF-DB02-40FC-BD8E-B11305970B98}.Release|Any CPU.Build.0 = Release|Any CPU
-		{C81E9F00-7899-4097-A7CE-043E818BE0E6}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
-		{C81E9F00-7899-4097-A7CE-043E818BE0E6}.Debug|Any CPU.Build.0 = Debug|Any CPU
-		{C81E9F00-7899-4097-A7CE-043E818BE0E6}.Release|Any CPU.ActiveCfg = Release|Any CPU
-		{C81E9F00-7899-4097-A7CE-043E818BE0E6}.Release|Any CPU.Build.0 = Release|Any CPU
+		{09079C11-2362-4FDE-ABD8-8309B75B55AE}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
+		{09079C11-2362-4FDE-ABD8-8309B75B55AE}.Debug|Any CPU.Build.0 = Debug|Any CPU
+		{09079C11-2362-4FDE-ABD8-8309B75B55AE}.Release|Any CPU.ActiveCfg = Release|Any CPU
+		{09079C11-2362-4FDE-ABD8-8309B75B55AE}.Release|Any CPU.Build.0 = Release|Any CPU
 	EndGlobalSection
 	GlobalSection(SolutionProperties) = preSolution
 		HideSolutionNode = FALSE
