+++ TradingSystem/TradingSystem/Controllers/SignUpInfo.cs
