+++ TradingSystem/TradingSystem/Controllers/WeatherForecastController.cs
