+++ TradingSystem/TradingSystem/Properties/launchSettings.json
