+++ TradingSystem/TradingSystem/Startup.cs
