+++ TradingSystem/TradingSystem/WeatherForecast.cs
