+++ TradingSystem/TradingSystem/appsettings.Development.json
