+++ TradingSystem/TradingSystem/appsettings.json
